requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
gast==0.3.2
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.24.3
h5py==2.10.0
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
imutils==0.5.3
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
moviepy==1.0.1
mtcnn==0.0.9
numpy==1.16.0
opencv-python==4.1.1.26
pandas==0.25.3
Pillow==6.2.1
proglog==0.1.9
protobuf==3.10.0
pydot==1.4.1
pyparsing==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.37.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
